fix(auth): fixed register unique constraints causing big red error on client side
